+++ packages/falcon-ecommerce-uikit/src/index.ts
@@ -28,5 +28,6 @@ export * from './Forms';
 export * from './AccountRecovery';
 export * from './Errors';
 export * from './SortOrdersProvider';
+export * from './OnlineStatus';
 export * from './EnsureTTI';
 export * from './BackendConfig';